Move Delim -> CTE rewrite to dedicated files
