+++ src/include/duckdb/planner/subquery/delim_join_cte_rewriter.hpp
@@ -1,0 +1,22 @@
+//===----------------------------------------------------------------------===//
+//                         DuckDB
+//
+// duckdb/planner/subquery/delim_join_cte_rewriter.hpp
+//
+//
+//===----------------------------------------------------------------------===//
+
+#pragma once
+
+#include "duckdb/planner/binder.hpp"
+#include "duckdb/planner/logical_operator.hpp"
+
+namespace duckdb {
+
+//! Rewrites fully decorrelated DelimJoins into materialized CTEs.
+class DelimJoinCTERewriter {
+public:
+	static void Rewrite(Binder &binder, unique_ptr<LogicalOperator> &plan);
+};
+
+} // namespace duckdb
